Fix some small issues with UI localization change from menu
